Finish refactor of cw-proposal-single tests.
